backend/auth-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
